back/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Flask-Mail==0.9.1
APScheduler==3.10.4
openai
pandas==2.2.2
openpyxl==3.1.2

